requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
arabic-reshaper==3.0.0
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.5.2
Pygments==2.19.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
google-genai
markdown